Initial Commit
We got here by:
mix phx.new victory --no-brunch --no-html --no-ecto

This is the starting point. Victory is yet to be determined.

In the moment when victory is achieved, it rarely looks glorious. It
looks tired, it looks beat down, it looks bloody, it looks like
tremendous loss. Victory is being the last one standing. Victory is
pushing through. There may be parades and confetti and statues later on,
but that is not victory - that is the celebration that came after the
hard work. The victory was achieved earlier, and it was there regardless
of whether or not celebration later occurred.

Live for victory.