All the way to Docker, via Makefile
This commit adds a Makefile. It has targets that do what you want. If it
doesn't do what you want, then you should fix it so it does, but before
you do that you should make sure you understand what it is doing. In
this case it is providing standard targets for things that I think you
would want to do. Things like cleaning, building, making a release, and
making a release inside a container.

I mentioned a container. It is built using a multi-stage Dockerfile,
which is the best thing since pour-over coffee, though not nearly as
old. This saves the massive headache of multiple docker files and
transfering artifacts between them. It is beautiful. The actual docker
file is completely generic. It would be great if we could keep it that
way, but probably we won't be able to. One can hope, though.